NOTIFY: Handle empty or multiple messages in a list
Summary: While there is no good reason to send a NOTIFY with an empty list of
message, at least HAProxy version 3.1+ had introduced a regression, now fixed in
https://github.com/haproxy/haproxy/commit/dc6e8dde231990d6f999c64117a301fc66a7256c,
which would cause empty NOTIFY to be sent.

e.g if using a spoe.cfg with the following:
```
spoe-message test_msg
    event on-frontend-http-request if { rand(100) gt 50 }
```

one would expect a NOTIFY to be sent 50% of the time, but in practice, a
NOTIFY is sent 100% of the time, with an empty list-of-message 50% of the time.
spop up to version 0.10.1 does not handle the empty list-of-message case, this
fixes it.

Likewise, multiple messages were not handled. This can be fixed by iterating
over the message parsing logic until the input is empty.
When doing so, the kvlist parser cannot expect to consume the whole input anymore.

Test Plan: Added test to cover empty message list, multiple messages.